Created Matrix class

+++ equation.h
@@ -5,9 +5,9 @@
 class Equation
 {
 private:
-    double x;
-    double y;
-    double c;
+    long double x;
+    long double y;
+    long double c;
 public:
     Equation(long double x, long double y, long double c);
     long double getX();
